Fix the memory leak
Though going by the Windows API documentation, it looks like an
allocation may occur even when the function fails. If that is
really true, then the function's projection in the windows crate
would seem to have a memory leak, separate from the one in this
code that this commit fixes.