Rename UnrecognizedOption to AttemptedOption
"UnrecognizedOption" was not a good name, because the point is that
no options at all are recognized. The goal here is to make clear
the distinction between flags, in the way that term is used in this
program, which supports them; and command-line options (which are
also called "flags" or, occasionally, "switches"), which are not
currently supported and which, if supported in the future, would
still presumably not themselves be convertible to KNOWN_FOLDER_FLAG
values.